Moved client form tests into AmendClientControllerTest from AmendClientTabControllerTest
