+++ src/test/java/uk/gov/justice/laa/amend/claim/controllers/amendments/AmendClientTabControllerTest.java
@@ -1,16 +1,11 @@
 package uk.gov.justice.laa.amend.claim.controllers.amendments;
 
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.request;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 import static uk.gov.justice.laa.amend.claim.utils.SessionUtils.AMENDMENTS_KEY;
 
-import java.util.Map;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -23,7 +18,7 @@ import uk.gov.justice.laa.amend.claim.models.AreaOfLaw;
 import uk.gov.justice.laa.amend.claim.models.ClaimDetails;
 import uk.gov.justice.laa.amend.claim.resources.MockClaimsFunctions;
 
-@WebMvcTest(controllers = {AmendClientTabController.class, AmendClientController.class})
+@WebMvcTest(controllers = {AmendClientTabController.class})
 class AmendClientTabControllerTest extends BaseControllerTest {
 
   private static final String INPUTS = "inputs[%s]";
@@ -49,84 +44,6 @@ class AmendClientTabControllerTest extends BaseControllerTest {
     claim.setClaimId(claimId);
     MockClaimsFunctions.updateStatus(claim, claim.getAssessmentOutcome());
     session.setAttribute(claimId.toString(), claim);
-  }
-
-  @Test
-  void savesFormsIntoSessionThenRedirectsClient1() throws Exception {
-    claim.setClientForename("forename");
-    claim.setClientSurname("surname");
-    claim.setClientGender("gender");
-    claim.setClientEthnicity("ethnicity");
-    claim.setClientDisability("disability");
-
-    var existingForms =
-        new AmendmentForms(new AmendmentForm(), new AmendmentForm(), new AmendmentForm());
-    session.setAttribute(AMENDMENTS_KEY.formatted(claimId), existingForms);
-
-    var client1Rows =
-        Map.of(
-            "INITIAL", FORENAME,
-            "SURNAME", SURNAME,
-            "GENDER", GENDER,
-            "ETHNICITY", ETHNICITY,
-            "DISABILITY", DISABILITY);
-    var client1Form = new AmendmentForm();
-    client1Form.setInputs(client1Rows);
-
-    var updatedForms =
-        new AmendmentForms(new AmendmentForm(), new AmendmentForm(), new AmendmentForm());
-    updatedForms.getClient1Form().setCurrent(client1Form);
-
-    var request = post(buildAmendClient1Path()).session(session).with(csrf());
-    for (var entry : client1Rows.entrySet()) {
-      request.param(INPUTS.formatted(entry.getKey()), entry.getValue());
-    }
-
-    mockMvc
-        .perform(request)
-        .andExpect(status().is3xxRedirection())
-        .andExpect(redirectedUrl(buildViewClientPath()))
-        .andExpect(request().sessionAttribute(AMENDMENTS_KEY.formatted(claimId), updatedForms));
-  }
-
-  @Test
-  void savesFormsIntoSessionThenRedirectsClient2() throws Exception {
-    claim.setClientForename("forename");
-    claim.setClientSurname("surname");
-    claim.setClientGender("gender");
-    claim.setClientEthnicity("ethnicity");
-    claim.setClientDisability("disability");
-
-    var existingForms =
-        new AmendmentForms(
-            new AmendmentForm(), new AmendmentForm(), new AmendmentForm(), new AmendmentForm());
-    session.setAttribute(AMENDMENTS_KEY.formatted(claimId), existingForms);
-
-    var client2Rows =
-        Map.of(
-            "CLIENT_2_INITIAL", FORENAME,
-            "CLIENT_2_SURNAME", SURNAME,
-            "CLIENT_2_GENDER", GENDER,
-            "CLIENT_2_ETHNICITY", ETHNICITY,
-            "CLIENT_2_DISABILITY", DISABILITY);
-    var client2Form = new AmendmentForm();
-    client2Form.setInputs(client2Rows);
-
-    var updatedForms =
-        new AmendmentForms(
-            new AmendmentForm(), new AmendmentForm(), new AmendmentForm(), new AmendmentForm());
-    updatedForms.getClient2Form().setCurrent(client2Form);
-
-    var request = post(buildAmendClient2Path()).session(session).with(csrf());
-    for (var entry : client2Rows.entrySet()) {
-      request.param(INPUTS.formatted(entry.getKey()), entry.getValue());
-    }
-
-    mockMvc
-        .perform(request)
-        .andExpect(status().is3xxRedirection())
-        .andExpect(redirectedUrl(buildViewClientPath()))
-        .andExpect(request().sessionAttribute(AMENDMENTS_KEY.formatted(claimId), updatedForms));
   }
 
   @Test
